requirements.txt
mysql-connector-python>=8.0.0
google-generativeai>=0.2.0
python-dotenv>=0.19.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-decouple>=3.8
PyPDF2>=3.0.0
pydantic>=2.0.0
python-json-logger>=2.0.0
langchain>=0.1.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.1
